fix: add suspense boundaries when using useSearchParams
