+++ next.config.mjs
@@ -9,9 +9,6 @@ const nextConfig = {
   images: {
     domains: ['github.com'],
   },
-  experimental: {
-    missingSuspenseWithCSRBailout: false,
-  },
 }
 
 export default nextConfig
