+++ src/app/(store)/search/loading.tsx
@@ -1,19 +1,15 @@
-'use client'
-
-import { useSearchParams } from 'next/navigation'
+import { Suspense } from 'react'
 
 import { Skeleton } from '@/components/skeleton'
 
+import { CurrentSearch } from './currentSearch'
+
 export default function SearchLoading() {
-  const searchParams = useSearchParams()
-
-  const query = searchParams.get('q')
-
   return (
     <div className="flex flex-col gap-4">
-      <p className="text-sm">
-        Resultados para: <span className="font-semibold">{query}</span>
-      </p>
+      <Suspense fallback={null}>
+        <CurrentSearch />
+      </Suspense>
       <div className="grid grid-cols-3 gap-6">
         <Skeleton className="h-[400px]" />
         <Skeleton className="h-[400px]" />
